rbt-web/requirements.txt
fastapi
uvicorn
pydantic
numpy
psycopg2-binary
redis
fastapi-cache2
pyjwt
